+++ src/SlowDiameter_3.java
@@ -1,5 +1,3 @@
-import Jcg.geometry.*;
-
 /**
  * Implementation of a quadratic time algorithm for computing a farthest pair of points (giving the diameter),
  * based on exhaustive search.
@@ -24,7 +22,7 @@ public class SlowDiameter_3 {
 		double distance=0.;
 		for(int i=0;i<points.length;i++) {
 			for(int j=i+1;j<points.length;j++) {
-				double pairDistance=points[i].distanceFrom(points[j]).doubleValue();
+				double pairDistance=points[i].distanceFrom(points[j]);
 				pair=new Point_3[] {points[i], points[j]};
 				if(pairDistance>distance) {
 					result=pair;
